gl/msaa: Add image "atlas" optimization.
In gl, we create a image atlas.  When an gl surface is to be painted to
target surface, we check whether the image can be cached into the atlas.
Using the image atlas, we can avoid flushing and rebinding textures of multiple
images if they are cached in the atlas.

For this version, we don't cache other backend surface yet.