NOT-FOR-UPSTREAM: gl: Don't check surface during context acquisition
Different drawable only means the gl context has different target window
or pixmap. In cairo case, we are OK because for offscreen rendering we
use FBO, for window or pixmap rendering, we always call _make_current.
In multithreaded application, different drivers behave differently.

For intel, the drawable is always 0 while context and display are not.
For nvidia, all three parameters are non-zero For fglrx, they are all 0.

FIXME: We can not upstream this patch until we can save/restore the
gl state properly.