gl: add GLESv2 support for ANGLE extension for multisampling
add support GL_ANGLE_blit_framebuffer and GL_ANGLE_multisample_framebuffer.
These two extensions are similar to what is available in GLESv2, so the
cairo implementation is also similar.  Basically, we always create
renderbuffer as rendering target regardless of whether the renderbuffer
is multisampling or not.  When the surface is used as mask or source,
we blit the content in renderbuffer to texture.